configured to run tests on TV
