+++ ui/wdio.tv.conf.js
@@ -1,5 +1,8 @@
 const ipAddress = require('../utils/ipAddress.js');
 const {config} = require('./wdio.conf.js');
+const {readChromeVersion} = require('../utils/readChromeVersionNumber');
+
+const chromeVersionMajorNumber = readChromeVersion();
 
 exports.config = Object.assign(
 	{},
@@ -9,14 +12,19 @@ exports.config = Object.assign(
 			maxInstances: 1,
 
 			browserName: 'chrome',
-			browserVersion: 'stable',
 			'goog:chromeOptions': {
 				w3c: false,
-				debuggerAddress: `${process.env.TV_IP}:9998`
+				debuggerAddress: '192.168.0.45:9998'
+			},
+			'wdio:chromedriverOptions': process.env.CHROME_DRIVER_PATH ? {
+				binary: process.env.CHROME_DRIVER_PATH
+				// match chromedriver version from jenkins
+			} : Number(chromeVersionMajorNumber) > 108 ? {} : {
+				binary: 'C:\\chromedriver\\chromedriver_v108.exe'
 			}
 		}],
 
-		baseUrl: `http://${ipAddress()}:4567`,
+		baseUrl: 'http://192.168.0.38:4567',
 
 		before: function () {
 			if (config.before) {
